+++ bin/check-mysql-status.rb
@@ -51,6 +51,11 @@ class CheckMySQLStatus < Sensu::Plugin::Check::CLI
          short: '-i',
          long: '--ini VALUE'
 
+  option :ini_section,
+         description: 'Section in my.cnf ini file',
+         long: '--ini-section VALUE',
+         default: 'client'
+
   option :hostname,
          description: 'Hostname to login to',
          short: '-h HOST',
@@ -107,7 +112,7 @@ class CheckMySQLStatus < Sensu::Plugin::Check::CLI
   def credentials
     if config[:ini]
       ini = IniFile.load(config[:ini])
-      section = ini['client']
+      section = ini[config[:ini_section]]
       db_user = section['user']
       db_pass = section['password']
       db_socket = section['socket']